+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Fixed
+
+- [TD-6895] Legacy df_content for concepts in elastic documents for rules and implementations
+
 ## [6.13.0] 2024-10-15
 
 ### Changed